+++ jdlib/secrets/commands/delete.py
@@ -1,0 +1,17 @@
+import sys
+from argparse import ArgumentParser
+from subprocess import CalledProcessError
+
+from jdlib.cli.commands import BaseCommand, ShellCommandMixin
+
+
+class Command(BaseCommand, ShellCommandMixin):
+    def add_arguments(self, parser: ArgumentParser):
+        parser.add_argument('name')
+
+    def handle(self, **options):
+        name = options['name']
+        try:
+            self.run_command(f'secret-tool clear collection jdlib name {name}')
+        except CalledProcessError:
+            sys.exit(1)