+++ jdlib/secrets/commands/create.py
@@ -1,8 +1,8 @@
 import getpass
-import sys
 from argparse import ArgumentParser
 from subprocess import CalledProcessError
 
+from jdlib.cli import call_command
 from jdlib.cli.commands import BaseCommand, ShellCommandMixin
 
 
@@ -12,8 +12,12 @@ class Command(BaseCommand, ShellCommandMixin):
 
     def handle(self, **options):
         name = options['name']
-        password = getpass.getpass('').strip()
-        try:
-            self.run_command(f'echo "{password}" | secret-tool store --collection="/org/freedesktop/secrets/collection/jdlib" --label="{name}" collection jdlib name {name}')
-        except CalledProcessError:
-            sys.exit(1)
+        result = call_command('secrets', 'get', name=name)
+        if result is None:
+            password = getpass.getpass('').strip()
+            try:
+                self.run_command(f'echo "{password}" | secret-tool store --collection="/org/freedesktop/secrets/collection/jdlib" --label="{name}" collection jdlib name {name}')
+            except CalledProcessError:
+                pass
+        else:
+            self.stdout.write(f'{name} already exists.')