+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## Unreleased
+## [0.0.7] - 2026-03-20
 
 ### Added
 
 - Manage secrets using `jdlib.secrets` module
 - `ShellCommandMixin` for executing shell commands
+- Call commands programatically using `call_command`
 
 ## [0.0.6] - 2026-03-20